requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyuic5-tool==0.0.1
